+++ modules/acm/variables.tf
@@ -90,5 +90,5 @@ variable "install_template_library" {
 variable "skip_gcloud_download" {
   description = "Whether to skip downloading gcloud (assumes gcloud and kubectl already available outside the module)"
   type        = bool
-  default     = false
+  default     = true
 }