+++ modules/asm/variables.tf
@@ -40,12 +40,6 @@ variable "skip_gcloud_download" {
   default     = true
 }
 
-variable "use_tf_google_credentials_env_var" {
-  description = "Optional GOOGLE_CREDENTIALS environment variable to be activated."
-  type        = bool
-  default     = false
-}
-
 variable "gcloud_sdk_version" {
   description = "The gcloud sdk version to use. Minimum required version is 293.0.0"
   type        = string
@@ -68,4 +62,22 @@ variable "gke_hub_membership_name" {
   description = "Memebership name that uniquely represents the cluster being registered on the Hub"
   type        = string
   default     = "gke-asm-membership"
+}
+
+variable "internal_ip" {
+  description = "Use internal ip for the cluster endpoint."
+  type        = bool
+  default     = false
+}
+
+variable "asm_dir" {
+  description = "Name of directory to keep ASM resource config files."
+  type        = string
+  default     = "asm-dir"
+}
+
+variable "asm_version" {
+  description = "ASM version to deploy. Available versions are documented in https://github.com/GoogleCloudPlatform/anthos-service-mesh-packages"
+  type        = string
+  default     = "release-1.6-asm"
 }