+++ modules/beta-private-cluster/versions.tf
@@ -15,9 +15,9 @@
  */
 
 terraform {
-  required_version = "~> 0.12.6"
+  required_version = ">=0.12.6, <0.14"
 
   required_providers {
-    google-beta = ">= 3.23.0, <4.0.0"
+    google-beta = ">= 3.29.0, <4.0.0"
   }
 }