+++ modules/beta-public-cluster/variables.tf
@@ -369,6 +369,7 @@ variable "default_max_pods_per_node" {
   default     = 110
 }
 
+
 variable "istio" {
   description = "(Beta) Enable Istio addon"
   default     = false
@@ -424,13 +425,10 @@ variable "enable_binary_authorization" {
   default     = false
 }
 
-variable "pod_security_policy_config" {
-  type        = list(object({ enabled = bool }))
+variable "enable_pod_security_policy" {
+  type        = bool
   description = "enabled - Enable the PodSecurityPolicy controller for this cluster. If enabled, pods must be valid under a PodSecurityPolicy to be created."
-
-  default = [{
-    "enabled" = false
-  }]
+  default     = false
 }
 
 variable "node_metadata" {