+++ modules/private-cluster/cluster.tf
@@ -195,7 +195,7 @@ resource "google_container_node_pool" "pools" {
 
   node_config {
     image_type   = lookup(each.value, "image_type", "COS")
-    machine_type = lookup(each.value, "machine_type", "n1-standard-2")
+    machine_type = lookup(each.value, "machine_type", "e2-medium")
     labels = merge(
       lookup(lookup(local.node_pools_labels, "default_values", {}), "cluster_name", true) ? { "cluster_name" = var.name } : {},
       lookup(lookup(local.node_pools_labels, "default_values", {}), "node_pool", true) ? { "node_pool" = each.value["name"] } : {},
@@ -264,7 +264,7 @@ resource "google_container_node_pool" "pools" {
 
 module "gcloud_wait_for_cluster" {
   source  = "terraform-google-modules/gcloud/google"
-  version = "~> 1.0.1"
+  version = "~> 1.3.0"
   enabled = var.skip_provisioners
 
   upgrade       = var.gcloud_upgrade