+++ modules/safer-cluster/variables.tf
@@ -328,13 +328,10 @@ variable "skip_provisioners" {
   default     = false
 }
 
-variable "pod_security_policy_config" {
-  type        = list(object({ enabled = bool }))
+variable "enable_pod_security_policy" {
+  type        = bool
   description = "enabled - Enable the PodSecurityPolicy controller for this cluster. If enabled, pods must be valid under a PodSecurityPolicy to be created."
-
-  default = [{
-    "enabled" = true
-  }]
+  default     = false
 }
 
 variable "gce_pd_csi_driver" {
@@ -359,4 +356,10 @@ variable "firewall_inbound_ports" {
   type        = list(string)
   description = "List of TCP ports for admission/webhook controllers"
   default     = ["8443", "9443", "15017"]
+}
+
+variable "config_connector" {
+  type        = bool
+  description = "(Beta) Whether ConfigConnector is enabled for this cluster."
+  default     = false
 }