+++ test/fixtures/sandbox_enabled/example.tf
@@ -32,8 +32,9 @@ module "example" {
 
   node_pools = [
     {
-      name       = "default-node-pool"
-      image_type = "COS_CONTAINERD"
+      name         = "default-node-pool"
+      image_type   = "COS_CONTAINERD"
+      machine_type = "n1-standard-2"
     },
   ]
 }